# contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Syncattend (SDAS) API Contract
  version: 0.1.0-draft
  description: >
    Shared API contract for Syncattend V3 (dual-auth smart attendance).
    OWNERSHIP: this file is the single source of truth for the HTTP contract.
    Only backend (owner A) may modify it. mobile (B) and web (C) READ ONLY.
    Layered defense: dynamic QR (time/identity) + ultrasonic audio token
    (space/proximity) + identity defense (1-min window + server uniqueness +
    app-generated UUID device binding + @wku.ac.kr re-registration) + professor-in-the-loop.
servers:
  - url: http://localhost:8000
    description: Local dev (backend)
...
paths:
...
  /auth/refresh:
    post:
      tags: [auth]
      summary: Refresh access token
      operationId: refresh
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refresh_token]
              properties:
                refresh_token: { type: string }
      responses:
        '200':
          description: New token pair
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TokenPair' }
        '401': { $ref: '#/components/responses/Unauthorized' }

  # ------------------------------------------------------------- devices
  /devices/register:
    post:
      tags: [devices]
      summary: Bind app-generated UUID to account (NO auto-bind for new/unknown UUID)
      description: >
        First binding of a new/unknown UUID must go through email re-auth.
        This endpoint binds only when the account has no binding conflict;
        otherwise it returns 409 and the client must use /devices/reregister/*.
      operationId: registerDevice
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [device_uuid]
              properties:
                device_uuid: { type: string, format: uuid }
      responses:
        '200':
          description: Device bound
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DeviceBinding' }
        '409':
          description: Binding conflict — email re-registration required
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /devices/reregister/request:
    post:
      tags: [devices]
      summary: Request device change — send verification code to @wku.ac.kr email
      operationId: reregisterRequest
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email:
                  type: string
                  format: email
                  description: Must be a @wku.ac.kr address
      responses:
        '202': { description: Verification code sent }
        '400': { $ref: '#/components/responses/BadRequest' }
  /devices/reregister/confirm:
    post:
      tags: [devices]
      summary: Confirm email code and rebind to new UUID
      operationId: reregisterConfirm
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, code, new_device_uuid]
              properties:
                email: { type: string, format: email }
                code: { type: string }
                new_device_uuid: { type: string, format: uuid }
      responses:
        '200':
          description: Rebound to new device
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DeviceBinding' }
        '400': { $ref: '#/components/responses/BadRequest' }
  /devices/me:
    get:
      tags: [devices]
      summary: Current account device binding
      operationId: getMyDevice
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Binding info
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DeviceBinding' }

  # ------------------------------------------------------------- courses
  /courses:
    post:
      tags: [courses]
      summary: Professor creates a course
      operationId: createCourse
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateCourseRequest' }
      responses:
        '201':
          description: Course created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Course' }
        '403': { $ref: '#/components/responses/Forbidden' }
    get:
      tags: [courses]
      summary: List courses owned by the authenticated professor
      operationId: listCourses
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Courses
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Course' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /courses/{course_id}/enrollments:
    post:
      tags: [courses]
      summary: Enroll a student in a course (professor-owned course only)
      operationId: enrollStudent
      security: [{ bearerAuth: [] }]
      parameters: [{ $ref: '#/components/parameters/CourseId' }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/EnrollRequest' }
      responses:
        '201':
          description: Student enrolled
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Enrollment' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409':
          description: Already enrolled
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
    get:
      tags: [courses]
      summary: List enrollments for a course (professor-owned course only)
      operationId: listEnrollments
      security: [{ bearerAuth: [] }]
      parameters: [{ $ref: '#/components/parameters/CourseId' }]
      responses:
        '200':
          description: Enrollments
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Enrollment' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }

  # ------------------------------------------------------------ sessions
...
  /sse/students/{id}:
    get:
      tags: [sse]
      summary: Student risk-group warning stream (SSE)
      operationId: sseStudent
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: text/event-stream of RiskWarning events
          content:
            text/event-stream:
              schema: { $ref: '#/components/schemas/RiskWarning' }
  /sse/sessions/{id}:
    get:
      tags: [sse]
      summary: Professor live attendance stream (SSE)
      operationId: sseSession
      security: [{ bearerAuth: [] }]
      parameters: [{ $ref: '#/components/parameters/SessionId' }]
      responses:
        '200':
          description: text/event-stream of AttendanceAggregate events
          content:
            text/event-stream:
              schema: { $ref: '#/components/schemas/AttendanceAggregate' }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    SessionId:
      name: id
      in: path
      required: true
      schema: { type: string }
    CourseId:
      name: course_id
      in: path
      required: true
      schema: { type: string }
  responses:
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    NotFound:
      description: Not found
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
  schemas:
...
    Error:
      type: object
      properties:
        detail: { type: string }
        code: { type: string }
...
    TokenPair:
      type: object
      properties:
        access_token: { type: string }
        refresh_token: { type: string }
        token_type: { type: string, example: bearer }
        role: { $ref: '#/components/schemas/Role' }
    Role:
      type: string
      enum: [student, professor]
...
    CreateCourseRequest:
      type: object
      required: [name]
      properties:
        name: { type: string }
    Course:
      type: object
      properties:
        id: { type: string }
        professor_id: { type: string }
        name: { type: string }
        created_at: { type: string, format: date-time }
    EnrollRequest:
      type: object
      required: [student_id]
      properties:
        student_id: { type: string }
    Enrollment:
      type: object
      properties:
        course_id: { type: string }
        student_id: { type: string }
    DeviceBinding:
      type: object
      properties:
        account_id: { type: string }
        device_uuid: { type: string, format: uuid }
        bound_at: { type: string, format: date-time }
...
    AttendanceAggregate:
      type: object
      properties:
        session_id: { type: string }
        total: { type: integer }
        present: { type: integer }
        unverified:
          type: array
          items: { type: string, description: student_id }
    RiskWarning:
      type: object
      properties:
        level: { type: string, enum: [info, warning, danger] }
        absences: { type: integer }
        message: { type: string }